+++ pkgs/tools/misc/rebuild/package.nix
@@ -2,7 +2,7 @@
 
 pkgs.writeShellScriptBin "rebuild" ''
   echo ${pkgs.stdenv.hostPlatform.system}
-  SYSTEM=${if pkgs.stdenv.isDarwin then "darwin" else "linux"}
+  SYSTEM=${if pkgs.stdenv.hostPlatform.isDarwin then "darwin" else "linux"}
   if [ "$SYSTEM" == "darwin" ]; then
       sudo darwin-rebuild switch --flake ${builtins.toString ../../../../.}
   else